# public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "0.0.1",
        "title": "Dwellar REST API"
    },
    "host": "52.74.34.255",
    "basePath": "/api/v1",
    "paths": {
...
        "/contacts": {
            "post": {
                "description": "This API allows creation of a contact.\n",
                "tags": [
                    "Contacts"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "access_token",
                        "in": "query",
                        "description": "Access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "photo",
                        "in": "query",
                        "description": "Photo of the contact. Should be in file type.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the Contact. Max 200 characters.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "birth_date",
                        "in": "query",
                        "description": "Date of birth. Format should be like 1995-06-01 (YYYY-mm-dd).",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Address of the contact",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "City Address of the contact",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "description": "State Address of the contact",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Country Address of the contact.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "zip",
                        "in": "query",
                        "description": "ZIP Code",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "Mobile Phones or telephone numbers of the contact",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email Address of the contact",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Tags of the contact. Should be comma separated values.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Text that describe the contact",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fb",
                        "in": "query",
                        "description": "Facebook Profile address of the contact",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "twitter",
                        "in": "query",
                        "description": "Twitter Profile address of the contact",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "google",
                        "in": "query",
                        "description": "Google+ Profile address of the contact",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "400": {
                        "description": "Error Response"
                    },
                    "401": {
                        "description": "Unauthorized Response",
                        "schema": {}
                    },
                    "403": {
                        "description": "Rate limit exceeded"
                    }
                }
            },
            "get": {
                "description": "This API lists all the contacts of the current logged user\n",
                "tags": [
                    "Contacts"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Returns the requested page. Each page can have up to 50 contacts",
                        "required": false,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "400": {
                        "description": "Error Response"
                    },
                    "403": {
                        "description": "Rate limit exceeded"
                    }
                }
            }
        },
        "/contacts/{contact_id}": {
            "get": {
                "description": "This API returns information for a single contact.\n",
                "tags": [
                    "Contacts"
                ],
                "parameters": [
                    {
                        "name": "contact_id",
                        "in": "path",
                        "description": "GUID of the contact for which the information is requested",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "400": {
                        "description": "Error Response"
                    },
                    "403": {
                        "description": "Rate limit exceeded"
                    },
                    "404": {
                        "description": "Error Response"
                    }
                }
            },
            "put": {
                "description": "This API updates a single existing contact.\n",
                "tags": [
                    "Contacts"
                ],
                "parameters": [
                    {
                        "name": "contact_id",
                        "in": "path",
                        "description": "GUID of the contact to update",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "photo",
                        "in": "query",
                        "description": "Photo of the contact. Should be in file type.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the Contact. Max 200 characters.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "birth_date",
                        "in": "query",
                        "description": "Date of birth. Format should be like 1995-06-01 (YYYY-mm-dd).",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Address of the contact",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "City Address of the contact",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "description": "State Address of the contact",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Country Address of the contact.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "zip",
                        "in": "query",
                        "description": "ZIP Code",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "Mobile Phones or telephone numbers of the contact",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email Address of the contact",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Tags of the contact. Should be comma separated values.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Text that describe the contact",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fb",
                        "in": "query",
                        "description": "Facebook Profile address of the contact",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "twitter",
                        "in": "query",
                        "description": "Twitter Profile address of the contact",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "google",
                        "in": "query",
                        "description": "Google+ Profile address of the contact",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "revision_number",
                        "in": "query",
                        "description": "Revision Number is required when updating contact. That way we can make sure that every data accross devices are correct",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "400": {
                        "description": "Error Response"
                    },
                    "403": {
                        "description": "Rate limit exceeded"
                    },
                    "404": {
                        "description": "Error Response"
                    }
                }
            },
            "delete": {
                "description": "This API deletes a single contact.\n",
                "tags": [
                    "Contacts"
                ],
                "parameters": [
                    {
                        "name": "contact_id",
                        "in": "path",
                        "description": "GUID of the contact for which the information is requested",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "400": {
                        "description": "Error Response"
                    },
                    "403": {
                        "description": "Rate limit exceeded"
                    },
                    "404": {
                        "description": "Error Response"
                    }
                }
            }
        },
...
    }
}